Fix line width/position precision loss far from origin at high zoom
At high zoom the camera sits very close to whatever point is being
viewed. If that point is far from the world origin, the view matrix's
translation is large, and computing (large view-space value) - (large
camera-position value) in float32 is a textbook catastrophic
cancellation: the result is placed correctly, but sub-pixel-scale math
built on top of it -- like the line-width perpendicular offset in
line-geometry.glsl -- loses so much precision it visibly thins out or
fades to a faint partial-coverage line. Confirmed by comparing
identical zoom levels near vs. far from the origin: near the origin
the line renders at full width, far from it the same line drops to a
sub-pixel gray line.

Fix by rendering camera-relative ("floating origin"): the view matrix
is now built relative to the camera's look-at point (m_center) instead
of true world coordinates, keeping its translation small and bounded
by cam_distance regardless of pan/zoom. Geometry that needs this
precision is shifted by -m_render_origin in double precision before
being narrowed to float for the GPU (Canvas::transform_point/_rel),
and draw_line/draw_axis_line now take double precision positions so
that shift happens before any float truncation, not after. Everything
that reads back through the view/projection matrices in the other
direction (project_to_window, get_cursor_pos_for_plane) is adjusted to
add the origin back. FaceGroup's per-group origin uniform is shifted
the same way so it stays in the same frame as the now-shifted vertex
data it's added to in face-vertex.glsl.

draw_point/draw_icon/draw_bitmap_text*/add_face_group's own inputs are
left at float precision for now -- narrower in scope, and not what the
reported symptom (line width shrinking at extreme zoom) touches.

+++ src/canvas/icanvas.hpp
@@ -46,8 +46,13 @@ public:
 
     virtual void clear() = 0;
     virtual VertexRef draw_point(glm::vec3 p) = 0;
-    virtual VertexRef draw_line(glm::vec3 from, glm::vec3 to) = 0;
-    virtual VertexRef draw_axis_line(glm::vec3 from, glm::vec3 to, Axis axis) = 0;
+    // Lines take double precision: at extreme zoom the camera sits close to
+    // geometry that may be far from the world origin, and float32 world
+    // coordinates lose enough precision in the view transform to visibly
+    // thin out (or misplace) the line-width offset computed in the
+    // geometry shader. See Canvas::transform_point/m_render_origin.
+    virtual VertexRef draw_line(glm::dvec3 from, glm::dvec3 to) = 0;
+    virtual VertexRef draw_axis_line(glm::dvec3 from, glm::dvec3 to, Axis axis) = 0;
     virtual VertexRef draw_screen_line(glm::vec3 origin, glm::vec3 direction) = 0;
     virtual std::vector<VertexRef> draw_bitmap_text(glm::vec3 p, float size, const std::string &rtext) = 0;
     virtual std::vector<VertexRef> draw_bitmap_text_centered(glm::vec3 p, float size,
